utils: move strlcpy/strlcat helpers from cgroup_utils to utils
