+++ src/cgroups/cgroup_utils.h
@@ -61,14 +61,6 @@ extern bool dir_exists(const char *path);
 extern int safe_mount(const char *src, const char *dest, const char *fstype,
 		      unsigned long flags, const void *data, const char *rootfs);
 
-#if !HAVE_STRLCPY
-extern size_t strlcpy(char *, const char *, size_t);
-#endif
-
-#if !HAVE_STRLCAT
-extern size_t strlcat(char *d, const char *s, size_t n);
-#endif
-
 extern FILE *fopen_cloexec(const char *path, const char *mode);
 extern void append_line(char **dest, size_t oldlen, char *new, size_t newlen);
 extern char *read_file(const char *fnam);
